+++ src/types/index.ts
@@ -15,20 +15,3 @@ export type Board = {
   title: string;
   lists: List[];
 };
-
-import type { DefaultSession } from "next-auth";
-import "next-auth/jwt";
-
-declare module "next-auth" {
-  interface Session {
-    user: {
-      id: string;
-    } & DefaultSession["user"];
-  }
-}
-
-declare module "next-auth/jwt" {
-  interface JWT {
-    id: string;
-  }
-}
